backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
Django==4.0.5
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
django-cors-headers
psycopg2-binary
dj-database-url
python-decouple
django-debreach
djoser==2.1.0
djangorestframework-simplejwt
drf-yasg
django-jazzmin
Pillow




